Resolve paths for grocery window
