+++ lib/main.js
@@ -4,6 +4,25 @@ const electron = require('electron')
 const app = electron.app
 // Module to create native browser window.
 const BrowserWindow = electron.BrowserWindow
+const windows = new Set()
+
+const createGroceryList = exports.createGroceryList = (file) => {
+  let groceryWindow = new BrowserWindow({show: false})
+  windows.add(groceryWindow)
+  
+  groceryWindow.loadURL(url.format({
+    pathname: path.join(__dirname, '../grocery-list.html'),
+    protocol: 'file:',
+    slashes: true
+  }))
+
+  groceryWindow.once('ready-to-show', () => {
+    if (file) openFile(groceryWindow, file)
+    groceryWindow.show()
+  })
+
+  return groceryWindow
+}
 
 const path = require('path')
 const url = require('url')
