+++ lib/renderer.js
@@ -1,9 +1,13 @@
-// This file is required by the index.html file and will
-// be executed in the renderer process for that window.
-// All of the Node.js APIs are available in this process.
-
+const electron = require('electron')
 const $ = require('jquery');
-const moment = require('moment');
+
+const ipc = electron.ipcRenderer
+const remote = electron.remote
+const shell = electron.shell
+
+const currentWindow = remote.getCurrentWindow()
+
+const { createGroceryList } = remote.require('./main')
 
 $('#save-recipe-btn').on('click', () => {
   const title = $('#title-field').val();
@@ -11,7 +15,6 @@ $('#save-recipe-btn').on('click', () => {
   const newLink = validateUrl(link);
   saveRecipe(title, newLink);
 })
-
 
 const saveRecipe = (title, link) => {
 
@@ -29,30 +32,6 @@ const validateUrl = (link) => {
   return link;
 }
 
-// Helper function, to format the time
-const secondsToTime = (s) => {
-  const momentTime = moment.duration(s, 'seconds');
-  const sec = momentTime.seconds() < 10 ? ('0' + momentTime.seconds()) : momentTime.seconds();
-  const min = momentTime.minutes() < 10 ? ('0' + momentTime.minutes()) : momentTime.minutes();
-  const hours = momentTime.hours() < 10 ? ('0' + momentTime.hours()) : momentTime.hours();
-  return `${hours}:${min}:${sec}`;
-}
-
-// Initialize currentTime
-let currentTime = 0;
-
-$('#increase-time').on('click', () => {
-  currentTime = currentTime + 10
-  timerDiv.innerHTML = secondsToTime(currentTime);
-})
-
-$('#start-timer').on('click', () => {
-  setInterval(() => {
-    // When reaching 0. Stop.
-    if(currentTime <= 0) return
-    // Remove one second
-    currentTime = currentTime - 1;
-    // Print out the time
-    timerDiv.innerHTML = secondsToTime(currentTime);
-  }, 1000); // 1 second
+$('#grocery-list-btn').on('click', (e) => {
+  createGroceryList()
 })
